input exceptions in create handled

+++ ieuRef.py
@@ -97,14 +97,13 @@ class MainWindow(QMainWindow):
         btn_createAuthorIdentityBibtex_action.setStatusTip('Create Author Identity')
         toolbar.addAction(btn_createAuthorIdentityBibtex_action)
 
-
         btn_deleteAllBibtex_action = QAction(QIcon("icon/delete.png"), "Delete All", self)
         btn_deleteAllBibtex_action.triggered.connect(self.deleteAllBibtex)
         btn_deleteAllBibtex_action.setStatusTip('Delete All')
         toolbar.addAction(btn_deleteAllBibtex_action)
 
     def loaddata(self, keyys):
-        MainWindow.data = MainWindow.data + keyys   # append(entry)
+        MainWindow.data = MainWindow.data + keyys  # append(entry)
         # newData = [(author, year, type1, title)]
         # entry = MainWindow.data
         # entry.append((author, year, type1, title))
@@ -131,7 +130,6 @@ class MainWindow(QMainWindow):
         #         print(column)
         #         self.tableWidget.setItem(row, column, QTableWidgetItem((data[row][column])))
 
-
     def createBibtex(self):
         dlg = CreateDialog()
         dlg.exec_()
@@ -141,9 +139,7 @@ class MainWindow(QMainWindow):
         # dlg.exec_()
         # if dlg.accepted():
 
-
         dlg.exec_()
-
 
     def filterBibtex(self):
         dlg = FilterDialog()
@@ -161,7 +157,7 @@ class MainWindow(QMainWindow):
         options = QFileDialog.Options()
         bibtexFile, _ = QFileDialog.getOpenFileName(self, "QFileDialog.getOpenFileName()", "",
                                                     "BibTeX Files (*.bib)", options=options)
-        if(bibtexFile):
+        if (bibtexFile):
             with open(bibtexFile) as bibtex:
 
                 try:
@@ -178,6 +174,7 @@ class MainWindow(QMainWindow):
                     QMessageBox.warning(QMessageBox(), 'Error', 'Could not load Bibtex file.')
         else:
             return
+
 
 class CreateDialog(QDialog):
     def __init__(self, *args, **kwargs):
@@ -259,14 +256,17 @@ class Article(QDialog):
         self.line5 = QLineEdit()
         self.line6 = QLineEdit()
 
-        self.text1 = QLabel("Author:")
-        self.text2 = QLabel("BibTeXKey:")
-        self.text3 = QLabel("Title:")
+        self.text1 = QLabel("Author(*):")
+        self.text2 = QLabel("BibTeXKey(*):")
+        self.text3 = QLabel("Title(*):")
         self.text4 = QLabel("Year:")
         self.text5 = QLabel("ID:")
         self.text6 = QLabel("Journal:")
+        self.keyText=QLabel("Key values must be filled(*)")
+        self.keyText.setStyleSheet('color: red')
 
         layout = QVBoxLayout()
+        layout.addWidget(self.keyText)
         layout.addWidget(self.text1)
         layout.addWidget(self.line1)
         layout.addWidget(self.text2)
@@ -285,12 +285,22 @@ class Article(QDialog):
         self.setLayout(layout)
 
     def btn_clk(self):
-        author = self.line1.text()
-        bibtexkey = self.line2.text()
-        title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
         journal = self.line6.text()
+        author = self.line1.text()
+        title = self.line3.text()
+        bibtexkey = self.line2.text()
+
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -345,9 +355,18 @@ class Book(QDialog):
         author = self.line1.text()
         bibtexkey = self.line2.text()
         title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
         publisher = self.line6.text()
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -398,8 +417,17 @@ class Journal(QDialog):
         author = self.line1.text()
         bibtexkey = self.line2.text()
         title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -436,7 +464,16 @@ class Proceeding(QDialog):
 
     def btn_clk(self):
         title = self.line1.text()
-        year = self.line2.text()
+        if title is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            year = int(self.line2.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open("Proceeding.bib", "w")
         newfile.write(
@@ -490,9 +527,18 @@ class InProceeding(QDialog):
         author = self.line1.text()
         bibtexkey = self.line2.text()
         title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
         booktitle = self.line6.text()
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -547,9 +593,18 @@ class MasterThesis(QDialog):
         author = self.line1.text()
         bibtexkey = self.line2.text()
         title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
         school = self.line6.text()
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -604,9 +659,19 @@ class PhdThesis(QDialog):
         author = self.line1.text()
         bibtexkey = self.line2.text()
         title = self.line3.text()
-        year = self.line4.text()
-        ID = self.line5.text()
         school = self.line6.text()
+
+        if author is "" or title is "" or bibtexkey is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
+        try:
+            ID = int(self.line5.text())
+            year = int(self.line4.text())
+
+        except ValueError:
+            QMessageBox.warning(QMessageBox(), 'Error', 'Year/ID must be a integer')
+            return
 
         newfile = open(str(bibtexkey) + ".bib", "w")
         newfile.write(
@@ -649,6 +714,10 @@ class Unpublished(QDialog):
         author = self.line1.text()
         title = self.line2.text()
         note = self.line3.text()
+        if author is "" or title is "":
+            QMessageBox.warning(QMessageBox(), 'Error', 'Key values can not be null')
+            return
+
 
         newfile = open("Unpublished.bib", "w")
         newfile.write(
@@ -676,13 +745,12 @@ class Misc(QDialog):
         self.setLayout(layout)
 
     def btn_clk(self):
-
-
         newfile = open("misc.bib", "w")
         newfile.write(
             "@" + "misc" + "{" "\n}")
 
         newfile.close()
+
 
 class FilterDialog(QDialog):
     def __init__(self, *args, **kwargs):
@@ -762,13 +830,13 @@ class FilterDialog(QDialog):
         self.label_4 = QtWidgets.QLabel(self.layoutWidget3)
         self.label_4.setObjectName("label_4")
         self.horizontalLayout_5.addWidget(self.label_4)
-        self.lineEdityearto = QtWidgets.QLineEdit(self.layoutWidget3)# year to
+        self.lineEdityearto = QtWidgets.QLineEdit(self.layoutWidget3)  # year to
         self.lineEdityearto.setObjectName("lineEdityearto")
         self.horizontalLayout_5.addWidget(self.lineEdityearto)
         self.label_5 = QtWidgets.QLabel(self.layoutWidget3)
         self.label_5.setObjectName("label_5")
         self.horizontalLayout_5.addWidget(self.label_5)
-        self.lineEditfrom = QtWidgets.QLineEdit(self.layoutWidget3)# year from
+        self.lineEditfrom = QtWidgets.QLineEdit(self.layoutWidget3)  # year from
         self.lineEditfrom.setObjectName("lineEditfrom")
         self.horizontalLayout_5.addWidget(self.lineEditfrom)
 
@@ -784,7 +852,6 @@ class FilterDialog(QDialog):
         self.pushButtonsearch.clicked.connect(self.saveTitletable)
         self.pushButtonsearch.clicked.connect(self.saveTypetabel)
         self.pushButtonsearch.clicked.connect(self.saveYear)
-
 
     def retranslateUi(self, Dialog):
         _translate = QtCore.QCoreApplication.translate
@@ -806,62 +873,61 @@ class FilterDialog(QDialog):
         self.label_4.setText(_translate("Dialog", "Year to"))
         self.label_5.setText(_translate("Dialog", "year from"))
 
-    def addNewRowAuthor(self):#add row Author table
+    def addNewRowAuthor(self):  # add row Author table
         rowPosition = self.tableWidgetauthor.rowCount()
         self.tableWidgetauthor.insertRow(rowPosition)
+
     def saveAuthortable(self):
         for row in range(self.tableWidgetauthor.rowCount()):
             rowdataauthor = []
             for column in range(self.tableWidgetauthor.columnCount()):
-                item = self.tableWidgetauthor.item(row,column)
+                item = self.tableWidgetauthor.item(row, column)
                 if item is not None:
 
                     rowdataauthor.append(item.text())
                 else:
                     rowdataauthor.append('')
 
-            print(rowdataauthor)#silincek
+            print(rowdataauthor)  # silincek
 
-
-
-    def addNewRowTitle(self):#add row Title table
-        rowPosition= self.tableWidgettitle.rowCount()
+    def addNewRowTitle(self):  # add row Title table
+        rowPosition = self.tableWidgettitle.rowCount()
         self.tableWidgettitle.insertRow(rowPosition)
+
     def saveTitletable(self):
         for row in range(self.tableWidgettitle.rowCount()):
             rowdatatitle = []
             for column in range(self.tableWidgettitle.columnCount()):
-                item = self.tableWidgettitle.item(row,column)
+                item = self.tableWidgettitle.item(row, column)
                 if item is not None:
 
                     rowdatatitle.append(item.text())
                 else:
                     rowdatatitle.append('')
 
-            print(rowdatatitle)#silinecek
+            print(rowdatatitle)  # silinecek
 
-
-    def addNewRowType(self):#add row Type table
-        rowPosition= self.tableWidgettype.rowCount()
+    def addNewRowType(self):  # add row Type table
+        rowPosition = self.tableWidgettype.rowCount()
         self.tableWidgettype.insertRow(rowPosition)
 
     def saveTypetabel(self):
         for row in range(self.tableWidgettype.rowCount()):
-            rowdatatype=[]
+            rowdatatype = []
             for column in range(self.tableWidgettype.columnCount()):
-                item = self.tableWidgettype.item(row,column)
+                item = self.tableWidgettype.item(row, column)
                 if item is not None:
 
                     rowdatatype.append(item.text())
                 else:
                     rowdatatype.append('')
 
-            print(rowdatatype)#silinecek
+            print(rowdatatype)  # silinecek
 
     def saveYear(self):
         yeardata = []
-        yearto=self.lineEdityearto.text()
-        yearfrom=self.lineEditfrom.text()
+        yearto = self.lineEdityearto.text()
+        yearfrom = self.lineEditfrom.text()
         print(yearto)
         print(yearfrom)
 
@@ -870,22 +936,16 @@ class FilterDialog(QDialog):
 
         if a < b:
             while a < b:
-                d = a+1
+                d = a + 1
                 yeardata.append(d)
                 a = d
         else:
             while b < a:
-                ey = b+1
+                ey = b + 1
                 yeardata.append(ey)
                 b = ey
 
-
-
-
-
-
         print(yeardata)
-
 
 
 class SearchDialog(QDialog):
@@ -1042,6 +1102,7 @@ class CreateAuthorIdentityDialog(QDialog):
                 for column, stuff in enumerate(row_data):
                     item = QTableWidgetItem(stuff)
                     self.tableWidget.setItem(row, column, item)
+
 
 app = QApplication(sys.argv)
 if (QDialog.Accepted == True):
